+++ State/PortfolioOperationEntry.cs
@@ -2,5 +2,5 @@ using System;
 
 namespace InvestmentAnalyzer.State {
 	public record PortfolioOperationEntry(
-		DateOnly Date, string BrokerName, string Type, string Currency, double Volume);
+		DateOnly Date, string BrokerName, string Type, string Currency, decimal Volume);
 }
